Split the Core\Error Controlller part to App\Controllers\Error
