+++ app/Controllers/Language.php
@@ -1,14 +1,26 @@
 <?php
 namespace App\Controllers;
 
+use Core\Controller;
 use Core\Language as CoreLanguage;
 use Helpers\Url;
 use Helpers\Cookie;
 use Helpers\Session;
 
 
-class Language
+class Language extends Controller
 {
+    /**
+     * Call the parent construct.
+     */
+    public function __construct()
+    {
+        parent::__construct();
+    }
+
+    /**
+     * Change the Framework Language.
+     */
     public function change($language)
     {
         // Only set language if it's in the Languages array
