+++ app/Views/Error/404.php
@@ -1,15 +1,10 @@
-<?php
-
-use Core\Error;
-
-?>
 <div class="container content">
 	<div class="row">
 		<div class="col-md-12">
 
 			<h1>404</h1>
 
-			<?php echo $data['error'];?>
+			<?= $error; ?>
 
 			<hr />
