+++ system/Core/Error.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Error class - calls a 404 page.
+ * Error class.
  *
  * @author David Carr - dave@novaframework.com
  * @version 3.0
@@ -8,47 +8,11 @@
 
 namespace Core;
 
-use Core\Controller;
-use Core\View;
-
 /**
- * Error class to generate 404 pages.
+ * Error class.
  */
-class Error extends Controller
+class Error
 {
-    /**
-     * $error holder.
-     *
-     * @var string
-     */
-    private $error = null;
-
-    /**
-     * Save error to $this->error.
-     *
-     * @param string $error
-     */
-    public function __construct($error = null)
-    {
-        parent::__construct();
-        $this->error = $error;
-    }
-
-    /**
-     * Load a 404 page with the error message.
-     */
-    public function index($error = null)
-    {
-        header("HTTP/1.0 404 Not Found");
-
-        $data['title'] = '404';
-        $data['error'] = $error ? $error : $this->error;
-
-        View::renderTemplate('header', $data);
-        View::render('Error/404', $data);
-        View::renderTemplate('footer', $data);
-    }
-
     /**
      * Display errors.
      *
